bug 1146430 - Move view_feature serializer code
The serialize code to support the view_feature viewset is twice the size
of all the other serializer code, and may change significantly w/ DRF
3.1.  Moved to own file.
